Use batch queue for logging of raw submission data

+++ smoon/process-batch-queue.py
@@ -18,6 +18,24 @@
 # along with this program; if not, write to the Free Software
 # Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301, USA.
 
+from optparse import OptionParser
+
+# Note: This stuff is up here so we don't get the warnings from inclusions
+#   when we don't even use that stuff, e.g. for --help
+parser = OptionParser(version = "sss")
+parser.add_option('--config',
+                  dest = 'config_file',
+                  default = None,
+                  metavar = 'file.cfg',
+                  help = 'override config file to use')
+parser.add_option('--delete',
+                  dest = 'delete_after_addition',
+                  default = False,
+                  action = 'store_true',
+                  help = 'delete entries after addition (default is marking as added)')
+
+(opts, args) = parser.parse_args()
+
 import sys
 import os
 import logging
@@ -33,23 +51,21 @@ from hardware.model.model import BatchJob
 # if it's not on the command line, then
 # look for setup.py in this directory. If it's not there, this script is
 # probably installed
-if len(sys.argv) > 1:
-    configfile = sys.argv[1]
-elif os.path.exists(os.path.join(os.path.dirname(__file__), "setup.py")):
-    configfile = 'dev.cfg'
-else:
-    configfile = 'prod.cfg'
+if opts.config_file == None:
+    if os.path.exists(os.path.join(os.path.dirname(__file__), "setup.py")):
+        opts.config_file = 'dev.cfg'
+    else:
+        opts.config_file = 'prod.cfg'
 
 
 config = ConfigParser()
-config.read(configfile)
+config.read(opts.config_file)
 CONNECTION = config.get('global', 'sqlalchemy.dburi').\
         lstrip('"\'').rstrip('"\'')
 engine = create_engine(CONNECTION, echo=True)
 session = sessionmaker(bind=engine)()
-
-count = session.query(BatchJob).order_by(BatchJob.arrival).count()
-jobs = session.query(BatchJob).order_by(BatchJob.arrival).all()
+count = session.query(BatchJob).filter_by(added=False).order_by(BatchJob.arrival).count()
+jobs = session.query(BatchJob).filter_by(added=False).order_by(BatchJob.arrival).all()
 for j in jobs:
     print '===================================================================='
     print 'Processing job with hardware UUID %s' % j.hw_uuid
@@ -58,7 +74,10 @@ for j in jobs:
         handle_submission(session, j.hw_uuid, j.data)
     except Exception, e:
         logging.debug('Caught exception: %s' % e.message)
-    session.delete(j)
+    if opts.delete_after_addition:
+        session.delete(j)
+    else:
+        j.added = True
 session.flush()
 print '===================================================================='
 print '%d jobs processed' % count
